fix: harden fake session read snapshots and clear PR review blockers
Deep-clone fake session state under read locks so lifecycle controls cannot race projection reads, remove the root progress.txt task artifact from the branch, and satisfy pkg-maint for the fake-service test surface.
